# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalInfluencers: number;
  totalBusinesses: number;
  totalCampaigns: number;
  totalPayoutsInr: number;
  avgEngagementRate: number;
}

export type InfluencerStatus = typeof InfluencerStatus[keyof typeof InfluencerStatus];


export const InfluencerStatus = {
  pending: 'pending',
  verified: 'verified',
} as const;

export interface Influencer {
  id: string;
  name: string;
  /** Primary platform, e.g. Instagram, YouTube, Facebook, X, LinkedIn, Snapchat */
  platform: string;
  /** e.g. Fashion, Tech, Fitness, Food, Travel, Finance */
  category: string;
  followers: number;
  /** Percentage engagement rate */
  engagementRate: number;
  location: string;
  pricePerPostInr: number;
  rating: number;
  bio: string;
  avatarUrl: string;
  status: InfluencerStatus;
  createdAt: string;
}

export interface CreateInfluencerInput {
  name: string;
  platform: string;
  category: string;
  followers: number;
  engagementRate?: number;
  location: string;
  pricePerPostInr: number;
  bio: string;
  avatarUrl?: string;
}

export type CampaignRequestStatus = typeof CampaignRequestStatus[keyof typeof CampaignRequestStatus];


export const CampaignRequestStatus = {
  new: 'new',
  matching: 'matching',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface CampaignRequest {
  id: string;
  businessName: string;
  contactEmail: string;
  contactPhone: string;
  productName: string;
  promotionType: string;
  targetAudience: string;
  budgetInr: number;
  durationDays: number;
  platforms: string[];
  deliverables: string;
  status: CampaignRequestStatus;
  createdAt: string;
}

export interface CreateCampaignInput {
  businessName: string;
  contactEmail: string;
  contactPhone: string;
  productName: string;
  promotionType: string;
  targetAudience: string;
  budgetInr: number;
  durationDays: number;
  platforms: string[];
  deliverables: string;
}

export type ListInfluencersParams = {
platform?: string;
category?: string;
search?: string;
};