# lib/api-zod/src/generated/types/campaignRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignRequestStatus } from './campaignRequestStatus';

export interface CampaignRequest {
  id: string;
  businessName: string;
  contactEmail: string;
  contactPhone: string;
  productName: string;
  promotionType: string;
  targetAudience: string;
  budgetInr: number;
  durationDays: number;
  platforms: string[];
  deliverables: string;
  status: CampaignRequestStatus;
  createdAt: Date;
}