# lib/api-zod/src/generated/types/createCampaignInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCampaignInput {
  businessName: string;
  contactEmail: string;
  contactPhone: string;
  productName: string;
  promotionType: string;
  targetAudience: string;
  budgetInr: number;
  durationDays: number;
  platforms: string[];
  deliverables: string;
}