# lib/api-zod/src/generated/types/createInfluencerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInfluencerInput {
  name: string;
  platform: string;
  category: string;
  followers: number;
  engagementRate?: number;
  location: string;
  pricePerPostInr: number;
  bio: string;
  avatarUrl?: string;
}