# lib/api-zod/src/generated/types/influencer.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InfluencerStatus } from './influencerStatus';

export interface Influencer {
  id: string;
  name: string;
  /** Primary platform, e.g. Instagram, YouTube, Facebook, X, LinkedIn, Snapchat */
  platform: string;
  /** e.g. Fashion, Tech, Fitness, Food, Travel, Finance */
  category: string;
  followers: number;
  /** Percentage engagement rate */
  engagementRate: number;
  location: string;
  pricePerPostInr: number;
  rating: number;
  bio: string;
  avatarUrl: string;
  status: InfluencerStatus;
  createdAt: Date;
}