# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalInfluencers: number;
  totalBusinesses: number;
  totalCampaigns: number;
  totalPayoutsInr: number;
  avgEngagementRate: number;
}